# back/central/shared/docs/modules/modules_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Modules service",
        "title": "Modules API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/payments/mappings": {
            "get": {
                "description": "Obtiene una lista de mapeos de métodos de pago con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Listar mapeos de métodos de pago",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por tipo de integración (shopify, whatsapp, mercadolibre)",
                        "name": "integration_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMappingsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo mapeo entre un método de pago original y uno unificado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Crear mapeo de método de pago",
                "parameters": [
                    {
                        "description": "Datos del mapeo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreatePaymentMappingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMappingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/mappings/integration/{type}": {
            "get": {
                "description": "Obtiene todos los mapeos de un tipo de integración específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Obtener mapeos por tipo de integración",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tipo de integración (shopify, whatsapp, mercadolibre)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.PaymentMappingResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/mappings/{id}": {
            "get": {
                "description": "Obtiene un mapeo de método de pago por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Obtener mapeo de método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del mapeo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMappingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un mapeo existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Actualizar mapeo de método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del mapeo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdatePaymentMappingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMappingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un mapeo de método de pago",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Eliminar mapeo de método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del mapeo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/mappings/{id}/toggle": {
            "patch": {
                "description": "Cambia el estado activo/inactivo de un mapeo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Mappings"
                ],
                "summary": "Activar/Desactivar mapeo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del mapeo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMappingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/methods": {
            "get": {
                "description": "Obtiene una lista paginada de métodos de pago con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Listar métodos de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Tamaño de página",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por categoría (card, digital_wallet, bank_transfer, cash)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscar por nombre o código",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMethodsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo método de pago en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Crear método de pago",
                "parameters": [
                    {
                        "description": "Datos del método de pago",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreatePaymentMethodRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/methods/{id}": {
            "get": {
                "description": "Obtiene un método de pago por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Obtener método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del método de pago",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un método de pago existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Actualizar método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del método de pago",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdatePaymentMethodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un método de pago (no se puede eliminar si tiene mapeos activos)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Eliminar método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del método de pago",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/methods/{id}/toggle": {
            "patch": {
                "description": "Cambia el estado activo/inactivo de un método de pago",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Activar/Desactivar método de pago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del método de pago",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreatePaymentMappingRequest": {
            "type": "object",
            "required": [
                "integration_type",
                "original_method",
                "payment_method_id"
            ],
            "properties": {
                "integration_type": {
                    "type": "string",
                    "enum": [
                        "shopify",
                        "whatsapp",
                        "mercadolibre"
                    ]
                },
                "original_method": {
                    "type": "string",
                    "maxLength": 128
                },
                "payment_method_id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "domain.CreatePaymentMethodRequest": {
            "type": "object",
            "required": [
                "category",
                "code",
                "name"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "enum": [
                        "card",
                        "digital_wallet",
                        "bank_transfer",
                        "cash"
                    ]
                },
                "code": {
                    "type": "string",
                    "maxLength": 64
                },
                "color": {
                    "type": "string",
                    "maxLength": 32
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "provider": {
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
        "domain.PaymentMappingResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "integration_type": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "original_method": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/domain.PaymentMethodResponse"
                },
                "payment_method_id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentMappingsListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PaymentMappingResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.PaymentMethodResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentMethodsListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PaymentMethodResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "domain.UpdatePaymentMappingRequest": {
            "type": "object",
            "required": [
                "original_method",
                "payment_method_id"
            ],
            "properties": {
                "original_method": {
                    "type": "string",
                    "maxLength": 128
                },
                "payment_method_id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "domain.UpdatePaymentMethodRequest": {
            "type": "object",
            "required": [
                "category",
                "name"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "enum": [
                        "card",
                        "digital_wallet",
                        "bank_transfer",
                        "cash"
                    ]
                },
                "color": {
                    "type": "string",
                    "maxLength": 32
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "provider": {
                    "type": "string",
                    "maxLength": 64
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}